fix: patch version handling
Factorio uses semantic versioning 2.0 for it's releases. When mods are
tied to releases the mod must specify the factorio release it should be
associated with down to the minor version. Apparently the

MAJOR.MINOR.PATCH

E.g.:

Factorio may be 1.1.18 and the mod may be released for version 1.1.107, 1.1.*,
or simply 1.1

If a mod does not specify its patch version it erroniously fails to
match and doesn't download. This causes the check to ensure the first
two components match and only proceeds to the second if specified.

Resolves #25